Speedup rank_sums implementation and use u32 1D array instead of i64 1D array as input.
Speedup ranks_sums implementation by presorting the y part of the concatenated x and y
array as presorting the y part (for which the original order does not need to be kept)
speeds up argsorting this array later (as argsort has a much bigger overhead than
plain sort). Also use u32 1D array instead of i64 array as input as that is the optimal
data type for the input data that later will be passed in the implementation.